Style: reponsive header, home - banner
