+++ package.json
@@ -63,6 +63,6 @@
     "@types/styled-components": "^5.1.34",
     "autoprefixer": "^10.4.16",
     "postcss": "^8.4.32",
-    "tailwindcss": "^3.3.6"
+    "tailwindcss": "^3.4.1"
   }
 }
