+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { RecoilRoot } from "recoil";
+import { HelmetProvider } from "react-helmet-async";
 
 import "./index.css";
 
@@ -31,12 +32,11 @@ import "./styles/globalStyle.css";
 import "./styles/applyEdu.css";
 
 import ScrollToTop from "./components/scrollToTop"; // 페이지 렌더링 시 스크롤 최상단
-import { Foot } from "./components/foot";
 import { Header } from "./components/Header";
+import { Footer } from "./components/Footer";
+import { Home } from "./pages/Home";
 import { NotFound } from "./pages/404";
 import { UpdatePost } from "./pages/post/UpdatePost";
-// import { EditPost } from "./pages/editPost";
-import { HomePage } from "./pages/home";
 import { Post } from "./pages/post/post";
 import { Posts } from "./pages/post/posts";
 import { ApplyEdu } from "./pages/applyEdu";
@@ -68,46 +68,48 @@ import AppPrivacyPolicy from "./pages/appPrivacyPolicy";
 export const App = () => {
   return (
     <BrowserRouter>
-      <RecoilRoot>
-        <ScrollToTop />
-        <Header />
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/post/updatepost" element={<UpdatePost />} />
-          <Route path="/posts" element={<Posts />} />
-          <Route path="/post" element={<Post />} />
-          <Route path="*" element={<NotFound />} />
-          <Route path="/applyEdu" element={<ApplyEdu />} />
-          <Route path="/update-application" element={<UpdateApplication />} />
-          <Route path="/showApplication" element={<ShowApplication />} />
-          <Route path="/showDetailContent" element={<ShowDetailContent />} />
-          <Route path="/login" element={<Login />} />
+      <HelmetProvider>
+        <RecoilRoot>
+          <ScrollToTop />
+          <Header />
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/post/updatepost" element={<UpdatePost />} />
+            <Route path="/posts" element={<Posts />} />
+            <Route path="/post" element={<Post />} />
+            <Route path="*" element={<NotFound />} />
+            <Route path="/applyEdu" element={<ApplyEdu />} />
+            <Route path="/update-application" element={<UpdateApplication />} />
+            <Route path="/showApplication" element={<ShowApplication />} />
+            <Route path="/showDetailContent" element={<ShowDetailContent />} />
+            <Route path="/login" element={<Login />} />
 
-          <Route path="/otto" element={<Otto />} />
-          <Route path="/hourGlass" element={<HourGlass />} />
-          <Route path="/spider" element={<Spider />} />
-          <Route path="/snake" element={<Snake />} />
-          <Route path="/smallPuppy" element={<SmallPuppy />} />
-          <Route path="/human" element={<Human />} />
-          <Route path="/security" element={<Security />} />
-          <Route path="/logistics" element={<Logistics />} />
-          <Route path="/securityDog" element={<SecurityDog />} />
-          <Route path="/yolo" element={<Yolo />} />
-          <Route path="/deepLearning" element={<DeepLearning />} />
-          <Route
-            path="/reinforcementLearning"
-            element={<ReinforcementLearning />}
-          />
-          <Route path="/printer" element={<Printer />} />
-          <Route path="/plc" element={<Plc />} />
-          <Route path="/kicad" element={<Kicad />} />
-          <Route path="/embedded" element={<Embedded />} />
-          <Route path="/ros2" element={<Ros2 />} />
-          <Route path="/manipulator" element={<Manipulator />} />
-          <Route path="/appPrivacyPolicy" element={<AppPrivacyPolicy />} />
-        </Routes>
-        <Foot />
-      </RecoilRoot>
+            <Route path="/otto" element={<Otto />} />
+            <Route path="/hourGlass" element={<HourGlass />} />
+            <Route path="/spider" element={<Spider />} />
+            <Route path="/snake" element={<Snake />} />
+            <Route path="/smallPuppy" element={<SmallPuppy />} />
+            <Route path="/human" element={<Human />} />
+            <Route path="/security" element={<Security />} />
+            <Route path="/logistics" element={<Logistics />} />
+            <Route path="/securityDog" element={<SecurityDog />} />
+            <Route path="/yolo" element={<Yolo />} />
+            <Route path="/deepLearning" element={<DeepLearning />} />
+            <Route
+              path="/reinforcementLearning"
+              element={<ReinforcementLearning />}
+            />
+            <Route path="/printer" element={<Printer />} />
+            <Route path="/plc" element={<Plc />} />
+            <Route path="/kicad" element={<Kicad />} />
+            <Route path="/embedded" element={<Embedded />} />
+            <Route path="/ros2" element={<Ros2 />} />
+            <Route path="/manipulator" element={<Manipulator />} />
+            <Route path="/appPrivacyPolicy" element={<AppPrivacyPolicy />} />
+          </Routes>
+          <Footer />
+        </RecoilRoot>
+      </HelmetProvider>
     </BrowserRouter>
   );
 };
