+++ src/components/Header.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
+import logoSymbolWhite from "../images/logo/logo-white.png";
 import logoSymbol from "../images/logo/logo-symbol.png";
 import { ReactComponent as LogoText } from "../images/logo/logo-text.svg";
 
@@ -14,6 +15,16 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 
 export function Header() {
+  const location = useLocation();
+  const position = location.pathname === "/" ? "absolute" : "";
+  const logoTextColor = location.pathname === "/" ? "#FFF" : "#333";
+  const logo = location.pathname === "/" ? logoSymbolWhite : logoSymbol;
+  const headerTextClass = `${
+    location.pathname === "/"
+      ? "text-white hover:text-[#faab66]"
+      : "text-rodu-grey hover:text-black"
+  } text-[18px] font-medium transition-colors duration-500`;
+
   const [isOpen, setIsOpen] = useState(false);
 
   const toggleMenu = () => {
@@ -22,42 +33,34 @@ export function Header() {
 
   return (
     <>
-      <header className="h-[70px] w-full px-6 sm:px-10 md:px-[100px] flex items-center justify-between fixed bg-white border-b border-[#e7e7e7] z-[49]">
-        <Link to="/" className="flex items-center gap-4">
-          <img src={logoSymbol} className="h-[28px]" alt="logo" />
-          <LogoText height={18} />
+      <header
+        className={`h-[42px] tablet:h-[120px] w-full px-[20px] tablet:px-[44px] lg:px-[120px] max-tablet:pt-6 flex items-center justify-between z-[49] ${position}`}
+      >
+        <Link to="/" className="flex items-center gap-[14px] tablet:gap-[16px]">
+          <img src={logo} className="h-[26px] tablet:h-[38px]" alt="logo" />
+          <LogoText className="h-[14px] tablet:h-[20px]" fill={logoTextColor} />
         </Link>
         <div className="md:flex hidden gap-9">
           <a
             href="https://smartstore.naver.com/robotedu"
             target="_blank"
             rel="noopener noreferrer"
-            className="text-rodu-grey text-[18px] font-medium hover:text-black transition-colors duration-500"
+            className={headerTextClass}
           >
             구매 링크
           </a>
-          <Link
-            to="/showDetailContent"
-            className="text-rodu-grey text-[18px] font-medium hover:text-black transition-colors duration-500"
-          >
+          <Link to="/showDetailContent" className={headerTextClass}>
             교육 콘텐츠
           </Link>
-          <Link
-            to="/applyEdu"
-            className="text-rodu-grey text-[18px] font-medium hover:text-black transition-colors duration-500"
-          >
+          <Link to="/applyEdu" className={headerTextClass}>
             교육 신청
           </Link>
-          <Link
-            to="/posts"
-            className="text-rodu-grey text-[18px] font-medium hover:text-black transition-colors duration-500"
-            aria-current="page"
-          >
+          <Link to="/posts" className={headerTextClass}>
             커뮤니티 게시판
           </Link>
         </div>
         <button onClick={toggleMenu} className="md:hidden">
-          <FontAwesomeIcon icon={faBars} className="text-black text-[24px]" />
+          <FontAwesomeIcon icon={faBars} className="text-white text-[22px]" />
         </button>
         {isOpen && (
           <div className="fixed top-0 left-0 h-full w-full bg-black bg-opacity-60 z-50">
