+++ src/types/custom.d.ts
@@ -1,0 +1,4 @@
+declare module "*.mp4" {
+  const src: string;
+  export default src;
+}
